* changes:
  Merge WebKit at r72805: Update WebKit revision
  Merge WebKit at r72805: Cherry-pick WebKit change 73062
  Merge WebKit at r72805: Cherry-pick WebKit change 72948
  Merge WebKit at r72805: Update call sites to RenderFrame::docWidth()/docHeight()
  Merge WebKit at r72805: Undo previous conflict resolution in RenderSVGResourceSolidColor.cpp
  Merge WebKit at r72805: Update GraphicsContext method signatures
  Merge WebKit at r72805: Update EditorClient::getGuessesForWord()
  Merge WebKit at r72805: Use Node::parent() rather than Node::parentNode()
  Merge WebKit at r72805: Fix Makefiles
  Merge WebKit at r72805: Initial merge by Git

+++ WebCore/xml/XMLHttpRequestProgressEventThrottle.cpp
@@ -47,7 +47,7 @@ XMLHttpRequestProgressEventThrottle::~XMLHttpRequestProgressEventThrottle()
 {
 }
 
-void XMLHttpRequestProgressEventThrottle::dispatchProgressEvent(bool lengthComputable, unsigned long long loaded, unsigned long long total)
+void XMLHttpRequestProgressEventThrottle::dispatchProgressEvent(bool lengthComputable, unsigned loaded, unsigned total)
 {
     ASSERT(!suspended());
     if (!isActive()) {
